Support queries that can't be run in a transaction
Before running queries, check whether any queries in the statement list cannot
be run in a transaction. If there are any, issue the statement directly against
the database instead of starting a transaction. Ensure that the statement block
only has one outside-of-transaction query before running it.

Adds tests for the new cannot-run-in-transaction query parser.