lib/goose,lib/goosedb: split into two packages
We carry a lot of dependencies - on mysql, postgres and sqlite3
drivers - which make the library import quite big. If all you want to
do is operate on the migrations directory, this can make the binary
size and import times prohibitive.

Split the package in two, with "goose" now focused on data structures
and migration parsing, and "goosedb" on actually interacting with a
database. Update commands in the binary to point to the new locations.

There is more that can be done here - we should do the file parsing in
"goose" instead of "goosedb" - but this is fine for now.